move functions from HCA helpers to platform
